requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-jazzmin==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.1
drf-yasg==1.21.10
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.6.0
